requirements.txt
agno
requests
python-dotenv